LayoutTests/platform/glib/imported/w3c/web-platform-tests/css/css-text/white-space/seg-break-transformation-016-expected.txt


FAIL linebreak only assert_equals: expected 83 but got 90
PASS zwsp retained 1
FAIL spaces linebreak assert_equals: expected 83 but got 90
PASS zwsp retained 2
FAIL linebreak spaces assert_equals: expected 83 but got 90
PASS zwsp retained 3
FAIL spaces linebreak spaces assert_equals: expected 83 but got 90
PASS zwsp retained 4
FAIL multiple linebreaks assert_equals: expected 83 but got 90
PASS zwsp retained 5
FAIL multiple linebreaks + spaces assert_equals: expected 83 but got 90
PASS zwsp retained 6
ภาษา​ ไทย
ภาษา​ ไทย
ภาษา​ ไทย
ภาษา​ ไทย
ภาษา​ ไทย
ภาษา​ ไทย
ภาษาไทย
